+++ docs/app/components/app/AppFooter.vue
@@ -11,9 +11,10 @@ const appConfig = useAppConfig()
       <footer>
         <div class="grid border-b border-default sm:grid-cols-2 lg:grid-cols-4">
           <div class="border-b border-default p-6 sm:border-r md:p-8 lg:border-b-0">
-            <AppHeaderLogo class="h-5 text-highlighted" />
+            <AppHeaderLogo class="h-5 w-auto text-highlighted" />
+
             <p class="mt-3 max-w-xs text-sm text-muted">
-              A fast, streaming-ready markdown parser with component support for Vue, React, and Svelte.
+              A fast, streaming-ready markdown parser with component support for Vue, React, Svelte, and Angular.
             </p>
           </div>